+++ include/main.hpp
@@ -3,21 +3,9 @@
 
 #include <Arduino.h>
 #include <string.h>
-
 #include "Context.hpp"
 #include "Messages.hpp"
-#include "Managers/Managers.hpp"
-
-struct RoboClawData {
-  uint8_t wheelSpeed;
-  bool valid;
-};
-
-void setupRoboClaw(Context &context);
-void setupNetwork(Context &context);
-void setupMotors(Context &context);
-void updateNetwork(Context &context);
-void updateRoboClaw(Context &context);
-void updateMotors(Context &context);
+#include "Managers/EthernetManager.hpp"
+#include "Managers/MotorManager.hpp"
 
 #endif
