+++ src/main.cpp
@@ -8,7 +8,6 @@ int main() {
   // }
   // Serial.println("Serial port connected to Teensy 4.1!");
 
-
   Context context;
   manager::MotorManager motorManager(context);
   manager::EthernetManager ethernetManager(context);
